controllers: create status-reporter job in operator ns
currently status-reporter job is created in the namespace
where storageclient gets created. To have a running pod
we need to have RBAC and servicAccount created on the
same namespace. We dont want to create the RBAC
from the code as it will get skipped in the security
audit and its a security problem. To solve it
when a storageclient gets created the status
reporter job will get created in the operator
namespace and it will get deleted when the
storageclient gets deleted.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>