controllers: derive operator Deployment name from Pod name
We were hardcoding the name of the operator's Deployment, which of
course breaks if the bundle generation is provided a different name
prefix for its resources. This implementation is still a bit of a hack
in that it implicitly relies on Kubernetes behaviors on how Pod names
are generated from Deployments... but better than nothing.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>