csi: Reduce memory alocation, GC, and "order of ops" considerations
The existing code is very wasteful regarding memory management of large
in-memory structures. On each reconcile cycle multiple big nested
structures are either fully or partially copied multiple times and in
multiple locations in the codebase. This creates unnecessary pressure
on the garbage collector and wastes multiple CPU cycles moving around
the same bits

A second side effect of the current code structure is the creation of
a fresh nested set of objects on every reconcile cycle that overwrites
any metadata on existing objects. This results in a contributor’s need
to be very careful and consider the order of operations when writing
reconciliation functions for CSI resources. As an example, I can give
the fact that the current code contains a bug where all of the owner
references on CSI reconciled objects are overwritten by a deep copy
operation.

This PR addresses the 2 points above and reduces the number of
allocations and garbage collection of in-memory reconciled objects and
the need to micro-manage the order of operation on CSI objects`
reconciliation.

Signed-off-by: nb-ohad <mitrani.ohad@gmail.com>