bundle: update missing and incorrect rbac resources
api resource name always is plural but we had incorrect rbac as `dns`
and it should be `dnses`. For other change, `rook-ceph-mon-endpoints` is
a name at provider end and in client cluster it is stored as
`ceph-csi-configs`.

The regression was introduced when we restricted the rbac to specific
resources at cluster role.

Signed-off-by: Leela Venkaiah G <lgangava@ibm.com>