controllers: sample changes for integrating ceph-csi operator
the intention is to start off with assuming ocs-client-operator deployed
with ceph-csi-operator need to be opt-in and as such made some changes.

worked with only available APIs so far and I'm afraid rewriting multiple
times the same pieces isn't going to work and so putting up this PR.

need a decision to decide on whether to provide a working code which can
work in two modes or else exclusively make it possible to work with new
operator only. the main hurdle will be refactory for each of these cases
which could result in a divergent code base which atm I'm unable to wrap
my head around.

Signed-off-by: Leela Venkaiah G <lgangava@ibm.com>