ci: refactor and fix gitignore & make targets
refactor of CI jobs:
divide the checks b/n code vs non-code which helps in easily turning off
one or the other during fixing any CI issues and also consolidates
jobs by similarity.

gitignore fix:
PR 311 brought new vendor dependency and that is being git ignored
unintentionally which created failures while running in CI, previous PR
was merged as not to get blocked by explicitly doing a `go mod tidy`
and `go mod vendor` which masked the original issue.

as not to hit this again, added a new condition for verifying deps from
mandatory parts for build if they are being git ignored.

golang cache:
as we were using `vendor/` for the project there is no need to cache mod
dependencies again and this is removed except for a single job where it
makes sense, infact storing & loading cache is increasing the overall
time w/ no added benefit.

concurrency of jobs:
added a concurrency check to the action as to cancel running jobs if PR
is pushed in quick successions as not to do work on older changes.

makefile:
do not update dependencies implicitiy before testing as that could mask
potentially issues which effectively makes updating dependencies a
separate step.

Signed-off-by: Leela Venkaiah G <lgangava@ibm.com>