controllers: always set metadata for csi nodeplugin pods
setting the option makes csi to add metadata to the volumes created in
ceph and until now this wasn't made default as there is a possibility
that external ceph couldn't support it.

however, based on available docs this is supported all the way from
RHCEPH 5 and current maintenance ODF version 4.15 supports RHCEPH 6.

Signed-off-by: Leela Venkaiah G <lgangava@ibm.com>