controllers: Set Consumer version to 4.18 until provider upgrade to 4.19
During an upgrade from 4.18 to 4.19, if the client side upgrades first,
the 4.18 provider will block its own upgrade upon detecting that the
client has a higher version. To avoid this deadlock, we intentionally
set the StorageConsumer's operatorVersion at 4.18 even after the
client upgrades to 4.19.

We detect whether the provider has upgraded to 4.19 using the
GetDesiredClientState RPC. This RPC is not available in 4.18, so if it
fails as Unimplemented, we assume the provider is still on 4.18 and
apply the 4.18 version to the StorageConsumer.

During the period we can not avoid updating the operatorVersion.
As if the operator version is empty in the Request it's set to N/A.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>